Admil
Added the NPC to steal like a romanian,
They walk to vault and if open they will steal, if not they'll be stalled.

Added the direction NPC finished finally, will ask for a direction as well as a random name and associated direction, No value for correct nor incorrect rewards.

Added a map to show all of these

Made the NPCs teleport to each level too

Billboard will be next and then the Bar NPCs